Profile singles cleanup, new output compares plotting script
